requirements.txt
discord.py
owoify-py
discord-py-slash-command